refactor: entity package를 model 패키지 내로 수정
